fix: deprecate np.int

+++ rankit/Table/Table.py
@@ -77,8 +77,8 @@ class Table(object):
         self.itemnum = idx
 
         # raw table need to be converted to standard indexed table.
-        raw_table['hidx'] = np.require(list(map(lambda x: itemlut[x], raw_table["host"].tolist())), dtype=np.int)
-        raw_table['vidx'] = np.require(list(map(lambda x: itemlut[x], raw_table["visit"].tolist())), dtype=np.int)
+        raw_table['hidx'] = np.require(list(map(lambda x: itemlut[x], raw_table["host"].tolist())), dtype=np.int32)
+        raw_table['vidx'] = np.require(list(map(lambda x: itemlut[x], raw_table["visit"].tolist())), dtype=np.int32)
 
         self.table = raw_table
